Unify SmartCache and add migration
