+++ ui/__init__.py
@@ -2,7 +2,8 @@
 from .design_tokens import DesignTokens
 from .modern_components import ModernButton, ModernCard, ModernProgressBar, ModernSidebar, LoadingIndicator, ModernTable
 from .style_manager import StyleSheetManager
-from .performance import SmartCache, AsyncTaskManager
+from .performance import AsyncTaskManager
+from optimizations.smart_cache import SmartCache
 
 __all__ = [
     'DesignTokens',
